Improve CTE materialization execution tests
Add methods checkCTEInfoMatch, verifyCTEExplainPlan and verifyResults
to test whether CTEs are materialized by checking if the Explain plan
has the expected CTEInfo element. The explain plan for CTEs was
updated to include isMaterialized info for CTEs when verbose optimizer
is enabled

Existing tests were only asserting if results of executing the query
are the same with and without using a CTE materialization strategy.
These new methods go further and also checks that the expected CTEs
were materialized per the Explain plan.

Refactor existing tests to use the verifyResults method or in the case
of more complex queries use the checkCTEInfoMatch method. In order for
the Explain plan to include the CTEInfo element, the system property
VERBOSE_OPTIMIZER_INFO_ENABLED needs to be set for any session
executing the query.